Routing 	-Add utils folder 	-Move Route paths to utils
