+++ src/utils/route_paths.js
@@ -1,0 +1,1 @@
+export const paths = { login: "/", notes: "/notes" };
